feat: implemented account deletion

+++ server/backend/controllers/userController.js
@@ -1,6 +1,7 @@
 const asyncHandler = require('express-async-handler')
 
 const User = require('../model/userModel')
+const Reservation = require('../model/reservationModel')
 
 // @desc    Get User
 // @route   GET /api/users
@@ -103,6 +104,7 @@ const updateUser = asyncHandler(async (req, res) => {
 // @route   DELETE /api/users/:id
 // @access  Private
 const deleteUser = asyncHandler(async (req, res) => {
+    const {password} = req.body;
     const user = await User.findById(req.params.id);
 
     if (!user) {
@@ -110,8 +112,25 @@ const deleteUser = asyncHandler(async (req, res) => {
         throw new Error('User not found');
     }
 
+    if (user.password !== password) {
+        res.status(401)
+        throw new Error('Password incorrect')
+    }
+
+    await Reservation.deleteMany({user: req.params.id});
     await user.deleteOne();
-    res.status(200).json({ id: req.params.id });
+
+    req.session.destroy((err) => {
+        if (err) {
+            res.status(500);
+            throw new Error('Could not log out after deletion')
+        }
+        res.clearCookie('connect.sid')
+        res.status(200).json({ 
+            message: 'User and all asssociated reservations deleted',
+            id: req.params.id
+        });
+    })
 })
 
 module.exports = {
